requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==4.2.7
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
drf-spectacular==0.27.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
phonenumbers==8.13.26
phonenumberslite==8.13.26
Pillow==10.1.0
psycopg2==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.16.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
